Refactor to allow easy support for various workloads
Move random sleep and hardcode numConc

Seed gofakeit in Setup so that all docs are different
